M2-5585: Add actions and formatRow hooks
