+++ src/modules/Dashboard/components/ActivityGrid/ActivityGrid.utils.tsx
@@ -11,7 +11,7 @@ export const getActivityActions = ({
   actions: { editActivity, exportData, assignActivity, takeNow },
   appletId,
   activityId,
-  dataTestid,
+  dataTestId,
   roles,
 }: ActivityActions): MenuItem<ActivityActionProps>[] => {
   const canEdit =
@@ -26,7 +26,7 @@ export const getActivityActions = ({
       title: t('editActivity'),
       context: { appletId, activityId },
       isDisplayed: canEdit,
-      'data-testid': `${dataTestid}-activity-edit`,
+      'data-testid': `${dataTestId}-activity-edit`,
     },
     {
       icon: <Svg id="export" />,
@@ -34,7 +34,7 @@ export const getActivityActions = ({
       title: t('exportData'),
       context: { appletId, activityId },
       isDisplayed: true,
-      'data-testid': `${dataTestid}-activity-export`,
+      'data-testid': `${dataTestId}-activity-export`,
     },
     { type: MenuItemType.Divider },
     {
@@ -43,7 +43,7 @@ export const getActivityActions = ({
       title: t('assignActivity'),
       context: { appletId, activityId },
       isDisplayed: true,
-      'data-testid': `${dataTestid}-activity-assign`,
+      'data-testid': `${dataTestId}-activity-assign`,
     },
     {
       icon: <Svg id="play-outline" />,
@@ -51,7 +51,7 @@ export const getActivityActions = ({
       title: t('takeNow'),
       context: { appletId, activityId },
       isDisplayed: true,
-      'data-testid': `${dataTestid}-activity-take-now`,
+      'data-testid': `${dataTestId}-activity-take-now`,
     },
   ];
 };
