+++ src/modules/Dashboard/components/ActivityGrid/index.ts
@@ -1,4 +1,5 @@
 export * from './ActivityGrid';
+export * from './ActivityGrid.hooks';
 export * from './ActivityGrid.styles';
 export * from './ActivityGrid.utils';
 export * from './ActivityGrid.types';
