+++ src/modules/Dashboard/features/Participant/Activities/Activities.tsx
@@ -1,33 +1,20 @@
-import { useState, useMemo, useEffect, useCallback } from 'react';
-import { generatePath, useNavigate, useParams } from 'react-router-dom';
-import { format } from 'date-fns';
+import { useState, useMemo, useEffect } from 'react';
+import { useParams } from 'react-router-dom';
 
-import { ActionsMenu, MenuActionProps, Row } from 'shared/components';
-import { StyledFlexTopCenter } from 'shared/styles';
 import { DatavizActivity, getSummaryActivitiesApi } from 'api';
 import { useAsync } from 'shared/hooks';
-import { DateFormats } from 'shared/consts';
-import { page } from 'resources';
-import { workspaces } from 'redux/modules';
-import {
-  ActivityGrid,
-  StyledSvg,
-  ActivityActionProps,
-  getActivityActions,
-} from 'modules/Dashboard/components/ActivityGrid';
+import { ActivityGrid, useActivityGrid } from 'modules/Dashboard/components/ActivityGrid';
 
 export const Activities = () => {
   const { appletId, participantId } = useParams();
-  const navigate = useNavigate();
   const [activitiesData, setActivitiesData] = useState<{
     result: DatavizActivity[];
     count: number;
   } | null>();
   const [isLoading, setIsLoading] = useState(true);
-  const dataTestid = 'dashboard-applet-participant-activities';
+  const dataTestId = 'dashboard-applet-participant-activities';
 
-  const workspaceRoles = workspaces.useRolesData();
-  const roles = appletId ? workspaceRoles?.data?.[appletId] : undefined;
+  const { formatRow } = useActivityGrid(dataTestId);
   /**
    * TODO M2-6223:
    * getAppletActivitiesApi returns activities for the currently logged in user (an admin/collaborator). This behavior is correct for M2-5585, as per the note:
@@ -58,100 +45,6 @@ export const Activities = () => {
     });
   }, [appletId, participantId, getSummaryActivities]);
 
-  const actions = useMemo(
-    () => ({
-      editActivity: ({ context }: MenuActionProps<ActivityActionProps>) => {
-        const { activityId } = context || {};
-
-        navigate(
-          generatePath(page.builderAppletActivity, {
-            appletId,
-            activityId,
-          }),
-        );
-      },
-      exportData: ({ context }: MenuActionProps<ActivityActionProps>) => {
-        const { activityId } = context || {};
-        // TODO: Implement export data
-        // https://mindlogger.atlassian.net/browse/M2-6039
-        alert(`TODO: Export data (${activityId})`);
-      },
-      assignActivity: ({ context }: MenuActionProps<ActivityActionProps>) => {
-        const { activityId } = context || {};
-        // TODO: Implement assign
-        // https://mindlogger.atlassian.net/browse/M2-5710
-        alert(`TODO: Assign activity (${activityId})`);
-      },
-      takeNow: ({ context }: MenuActionProps<ActivityActionProps>) => {
-        // TODO: Implement Take Now
-        // https://mindlogger.atlassian.net/browse/M2-5711
-        const { activityId } = context || {};
-        alert(`TODO: Take now (${activityId})`);
-      },
-    }),
-    [appletId, navigate],
-  );
-
-  const formatRow = useCallback(
-    (activity: DatavizActivity): Row => {
-      const activityId = String(activity.id);
-      const name = activity.name;
-      // TODO: getSummaryActivitiesApi needs to be updated to return `image`
-      const image = String('image' in activity ? activity.image : '');
-
-      // TODO: Populate with data from BE when available
-      // (Jira ticket(s) still being drafted)
-      const participantCount: number | null = null;
-      const compliance: number | null = null;
-      const trending: 'up' | 'down' | null = null;
-      const latestActivity: string | null = null;
-
-      return {
-        image: {
-          content: () => image,
-          value: image,
-        },
-        name: {
-          content: () => name,
-          value: name,
-        },
-        participantCount: {
-          content: () => participantCount,
-          value: Number(participantCount),
-        },
-        latestActivity: {
-          content: () =>
-            !!latestActivity &&
-            format(new Date(String(latestActivity)), DateFormats.MonthDayYearTime),
-          value: String(latestActivity),
-        },
-        compliance: {
-          content: () =>
-            typeof compliance === 'number' && (
-              <StyledFlexTopCenter as="span" sx={{ gap: 0.4 }}>
-                {compliance}% {!!trending && <StyledSvg id={`trending-${trending}`} />}
-              </StyledFlexTopCenter>
-            ),
-          value: Number(compliance),
-        },
-        actions: {
-          content: () =>
-            !!appletId && (
-              <ActionsMenu
-                menuItems={getActivityActions({ actions, appletId, activityId, dataTestid, roles })}
-                data-testid={`${dataTestid}-activity-actions`}
-                buttonColor="secondary"
-                anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-                transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-              />
-            ),
-          value: '',
-        },
-      };
-    },
-    [actions, appletId, roles],
-  );
-
   const activities = useMemo(
     () => (activitiesData?.result ?? []).map((activity) => formatRow(activity)),
     [activitiesData, formatRow],
@@ -163,7 +56,7 @@ export const Activities = () => {
       order={'desc'}
       orderBy={''}
       isLoading={isLoading}
-      data-testid={dataTestid}
+      data-testid={dataTestId}
     />
   );
 };
